changed controllers's structure
