+++ app/controllers/api/v1/books_controller.rb
@@ -1,0 +1,15 @@
+module Api
+  module V1
+    class BooksController < ApiController
+      def index
+        @books = Book.all
+        render_paginated @books
+      end
+
+      def show
+        @book = Book.find(params[:id])
+        render json: @book
+      end
+    end
+  end
+end
