+++ app/controllers/application_controller.rb
@@ -1,4 +1,3 @@
 # Default controller created by Rails
 class ApplicationController < ActionController::Base
-  
 end
